deploy/fetch_pool_logs_deploy/requirements.txt
# Function dependencies, for example:
# package>=version
threading
time
textwarp
pandas>=1.0.4
numpy==1.23.3
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
google-cloud==0.34.0
google-cloud-secret-manager==2.0.0
json==2.0.9
oauth2client==4.1.3
oauthlib==3.1.0
pandas-gbq>=0.14.0
requests>=2.28.1